backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
django-filter==21.1
psycopg2==2.8.6
Pillow==8.3.1